Crate rustc_unicode

Unstable (unicode #27783)

The Unicode Library

Unicode-intensive functions for char and str types.

This crate provides a collection of Unicode-related functionality, including decompositions, conversions, etc., and provides traits implementing these functions for the char and str types.

The functionality included here is only that which is necessary to provide for basic string-related manipulations. This crate does not (yet) aim to provide a full set of Unicode tables.

Modules

char
A character type.

derived_property 	 [Unstable]

str 	 [Unstable]