Module rustc_unicode::str

Unstable (unicode #27783)

Structs

SplitWhitespace
An iterator over the non-whitespace substrings of a string, separated by any amount of whitespace.

Utf16Encoder 	 [Unstable]
Iterator adaptor for encoding chars to UTF-16.

Utf16Items 	 [Deprecated]
An iterator that decodes UTF-16 encoded codepoints from a vector of u16s.

Enums

Utf16Item 	 [Deprecated]
The possibilities for values decoded from a u16 stream.

Traits

UnicodeStr 	 [Unstable]
Methods for Unicode string slices

Functions

is_utf16 	 [Unstable]
Determines if a vector of u16 contains valid UTF-16

utf16_items 	 [Deprecated]
Create an iterator over the UTF-16 encoded codepoints in v, returning invalid surrogates as LoneSurrogates.

utf8_char_width 	 [Unstable]
Given a first byte, determine how many bytes are in this UTF-8 character